deleted userinput.py
